+++ app/controllers/admin/index.js
@@ -1,0 +1,22 @@
+'use strict';
+
+var admin = require('./admin');
+var article = require('./article');
+var carousel = require('./carousel');
+var cate = require('./cate');
+var login = require('./login');
+var picture = require('./picture');
+var tag = require('./tag');
+var user = require('./user');
+
+
+// 暴露
+exports.admin = admin;
+exports.article = article;
+exports.carousel = carousel;
+exports.cate = cate;
+exports.login = login;
+exports.picture = picture;
+exports.tag = tag;
+exports.user = user;
+