+++ app/controllers/home/home.js
@@ -5,15 +5,6 @@ var mongoose = require('mongoose'),
   Article = mongoose.model('Article'),
   Cate = mongoose.model('Cate');
 
-exports.index = function (req, res) {
-    res.render('home/index', {
-      title: 'TheOne记事本|theone.io|吴斌杰|js工程师',
-      description : '个人笔记本或个人博客,总结个人工作的经验或学习经验的记录,同时实践ng,nodejs,bootstrap前后端一种语言javascript',
-      keywords: 'theone,theone.io,吴斌杰,js工程师,angular,nodejs,php,css3'
-    });
-    res.end();
-};
-
 //  文章列表
 exports.list = function (req, res, next) {
   var listCate = req.params.cate,