+++ app/controllers/home/index.js
@@ -1,0 +1,12 @@
+'use strict';
+
+var home = require('./home');
+var article = require('../article');
+var carousel = require('../carousel');
+var search = require('../search');
+
+exports.home = home;
+exports.article = article;
+exports.carousel = carousel;
+exports.search = search;
+